% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo.R
\name{exo_query}
\alias{exo_query}
\title{Run an ADQL query against the Exoplanet Archive TAP service}
\usage{
exo_query(sql, format = "json")
}
\arguments{
\item{sql}{ADQL query string (e.g. "select pl_name, hostname from ps
where disc_year = 2024")}

\item{format}{Output format: "json" (default), "csv", "votable"}
}
\value{
tibble with query results
}
\description{
Run an ADQL query against the Exoplanet Archive TAP service
}
